requirements.txt
discord.py==2.6.4
yt-dlp==2024.12.23
PyNaCl==1.5.0
spotipy==2.24.0
